# openapi.yaml
openapi: 3.0.3
...
info:
  title: RidgeRiceTalk API
  description: |
    RidgeRiceTalk — 自托管语音协作空间 API。

    设计原则：
    - 整个服务器即唯一空间（单例模式，无多租户）
    - 面向朋友/小团队，权限简单（OWNER/ADMIN/MEMBER）
    - 语音为核心，其他功能为辅助
    - 部署简单，SQLite 零配置默认

    功能模块：即时消息、语音通话（LiveKit）、音乐机器人、文件共享、
    白板协作、屏幕共享、小游戏、日程安排。

    认证方式：Bearer Token (JWT) + Session Cookie 双模式。
  version: 1.0.0
  contact:
    name: RidgeRiceTalk API Team
servers:
- url: https://api.ridgericetalk.example.com/api
  description: 生产环境
- url: http://localhost:8080/api
  description: 本地开发环境
security:
- BearerAuth: []
- CookieAuth: []

# 全局标签定义
...
paths:
...
  /api/v1/preview:
    get:
      tags: [Preview]
      summary: URL链接预览
      description: 获取指定URL的元数据信息（标题、描述、图片），用于消息中的链接预览卡片。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: url
        in: query
        required: true
        schema:
          type: string
          format: uri
        description: 要预览的URL
        example: https://www.example.com/article/123
      responses:
        '200':
          description: 预览信息
          content:
            application/json:
              schema:
                type: object
                required:
                - url
                - title
                - description
                - imageUrl
                properties:
                  url:
                    type: string
                    format: uri
                    description: 原始URL
                    example: https://www.example.com/article/123
                  title:
                    type: string
                    description: 页面标题
                    example: 如何构建高效的技术团队
                  description:
                    type: string
                    description: 页面描述
                    example: 本文分享了构建高效技术团队的实践经验和方法论...
                  imageUrl:
                    type: string
                    format: uri
                    nullable: true
                    description: 页面图片URL
                    example: https://www.example.com/images/team.jpg
        '400':
          description: URL无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                PREVIEWInvalidUrl:
                  value:
                    code: PREVIEW_INVALID_URL
                    message: URL格式无效
                PREVIEWFetchFailed:
                  value:
                    code: PREVIEW_FETCH_FAILED
                    message: 无法获取页面内容
...
  /api/v1/cloudfs/list:
    get:
      tags:
      - CloudFS
      summary: 列出文件/文件夹
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: path
        in: query
        schema:
          type: string
          default: /
      responses:
        '200':
          description: 目录内容
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CloudFileItem'
                  path:
                    type: string
  /api/v1/cloudfs/folder:
    post:
      tags:
      - CloudFS
      summary: 创建文件夹
      security:
      - BearerAuth: []
      - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - path
              - name
              properties:
                path:
                  type: string
                name:
                  type: string
      responses:
        '200':
          description: 创建成功
        '409':
          description: 文件夹已存在
  /api/v1/cloudfs/upload:
    post:
      tags:
      - CloudFS
      summary: 上传文件
      security:
      - BearerAuth: []
      - CookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                path:
                  type: string
                  default: /
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传成功
        '413':
          description: 文件过大
  /api/v1/cloudfs/download:
    get:
      tags:
      - CloudFS
      summary: 下载文件
      description: |
        以 attachment 返回文件流。若该文件上传时已记录完整性校验值，响应会带
        X-File-Checksum: sha256=<lowercase-hex>，客户端可据此校验下载内容；
        历史文件无该校验值时不返回该响应头。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: path
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 文件流
          headers:
            X-File-Checksum:
              schema:
                type: string
              description: 'sha256=<hex>；仅当存在已记录的校验值时返回'
        '404':
          description: 文件不存在
  /api/v1/cloudfs/delete:
    delete:
      tags:
      - CloudFS
      summary: 删除文件或文件夹
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: path
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 删除成功
        '404':
          description: 项目不存在
  /api/v1/cloudfs/preview:
    get:
      tags:
      - CloudFS
      summary: 在线预览文件
      description: |
        与 download 同鉴权、同权限校验（Space 成员 + 文件可见性），但按 MIME 白名单
        以内联方式返回（Content-Disposition: inline），便于图片/文本/PDF/音视频
        直接渲染。白名单之外（含 image/svg+xml、text/html）一律回落为 attachment，
        且 Content-Type 降级为 application/octet-stream + X-Content-Type-Options: nosniff。
        支持 HTTP Range（206），可分段加载大文件。download 端点行为不变。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: path
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 文件流（白名单类型为 inline，其余为 attachment）
          headers:
            Content-Disposition:
              schema:
                type: string
              description: inline（白名单类型）或 attachment（回落）
            X-Content-Type-Options:
              schema:
                type: string
              description: 固定为 nosniff
            X-File-Checksum:
              schema:
                type: string
              description: 'sha256=<hex>；仅当存在已记录的校验值时返回'
        '206':
          description: 分段内容（Range 请求）
        '400':
          description: 缺少 path 参数
        '403':
          description: 非 Space 成员或无权限
        '404':
          description: 文件不存在
  /api/v1/cloudfs/rename:
    patch:
      tags:
      - CloudFS
      summary: 重命名文件或文件夹
      description: |
        在同一父目录内改名；文件本体（物理存储）不动，仅更新逻辑名与路径。
        文件夹改名会递归更新整棵子树的 path / file_path（单事务）。
        权限与删除一致：上传者/所有者本人或 Space 管理员。
      security:
      - BearerAuth: []
      - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - path
              - newName
              properties:
                path:
                  type: string
                  description: 现有文件或文件夹的完整路径
                newName:
                  type: string
                  description: 新名称（单层路径段，不含 /）
      responses:
        '200':
          description: 重命名成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      type:
                        type: string
                        enum: [file, folder]
                      path:
                        type: string
                      name:
                        type: string
        '400':
          description: 名称为空/非法（CLOUDFS_NAME_INVALID）或过长（CLOUDFS_NAME_TOO_LONG）
        '403':
          description: 无权限（CLOUDFS_PERMISSION_DENIED）
        '404':
          description: 文件或文件夹不存在
        '409':
          description: 目标名称已存在（CLOUDFS_ALREADY_EXISTS）
  /api/v1/cloudfs/move:
    post:
      tags:
      - CloudFS
      summary: 移动文件或文件夹
      description: |
        移入目标父目录并保留原名。文件更新 folder_id 与 file_path；文件夹更新
        parent_id 并递归改写子树 path / file_path（单事务，任一目标路径冲突则整体回滚）。
        不允许移动到自身或其子目录。权限与删除一致，另需目标目录写权限。
      security:
      - BearerAuth: []
      - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - path
              - targetPath
              properties:
                path:
                  type: string
                  description: 现有文件或文件夹的完整路径
                targetPath:
                  type: string
                  description: 目标父目录路径，根目录为 /
      responses:
        '200':
          description: 移动成功
        '400':
          description: 目标目录不存在、或移动到自身/子目录
        '403':
          description: 无权限（CLOUDFS_PERMISSION_DENIED）
        '404':
          description: 文件或文件夹不存在
        '409':
          description: 目标位置同名项已存在（CLOUDFS_ALREADY_EXISTS）
  /api/v1/cloudfs/search:
    get:
      tags:
      - CloudFS
      summary: 按文件名搜索
      description: |
        按文件名模糊搜索（大小写不敏感，`%`/`_` 按字面量匹配，不做通配）。
        默认范围为整个 Space；传入 path 时限定在该目录及其子树内（递归）。
        仅返回文件（不含文件夹），最多 100 条，按创建时间倒序。
        权限与 list 一致：Space 成员即可，不额外做条目级鉴权。
        已进入回收站（软删）的条目不参与搜索。结果含 path 字段，便于跨目录定位。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: q
        in: query
        required: true
        description: 文件名关键词（不能为空）
        schema:
          type: string
      - name: path
        in: query
        required: false
        description: 限定搜索的目录路径（含其子树）；缺省为整个 Space
        schema:
          type: string
      responses:
        '200':
          description: 搜索结果（最多 100 条）
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      query:
                        type: string
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            type:
                              type: string
                              enum: [file]
                            path:
                              type: string
                              description: 命中文件的完整逻辑路径
                            size:
                              type: string
                            uploadBy:
                              type: string
                            updatedAt:
                              type: string
        '400':
          description: 缺少或为空的 q
        '403':
          description: 非 Space 成员
  /api/v1/cloudfs/trash:
    get:
      tags:
      - CloudFS
      summary: 列出回收站（已删除项）
      description: |
        列出当前 Space 内被删除（软删）的顶层条目，最新的在前。删除文件夹时整棵子树
        随文件夹一起进入回收站，其子项不再单独列出（与系统回收站一致）。
        回收站内容不计入 /cloudfs/usage 的用量统计，但在保留期内仍占用磁盘。
      security:
      - BearerAuth: []
      - CookieAuth: []
      responses:
        '200':
          description: 回收站列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      retainDays:
                        type: integer
                        description: 保留天数（过期后由后台任务彻底删除）
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            type:
                              type: string
                              enum: [file, folder]
                            path:
                              type: string
                            size:
                              type: string
                            uploadBy:
                              type: string
                            deletedAt:
                              type: string
                            expiresAt:
                              type: string
        '403':
          description: 非 Space 成员
    delete:
      tags:
      - CloudFS
      summary: 彻底删除回收站中的单项
      description: 删除 DB 行、FileMetadata 行与磁盘内容，不可恢复。文件夹会连同其已删除子树一并清除。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: id
        in: query
        required: false
        description: 条目 ID（与 path 二选一）
        schema:
          type: string
      - name: path
        in: query
        required: false
        description: 条目路径（与 id 二选一）
        schema:
          type: string
      responses:
        '200':
          description: 已彻底删除
        '400':
          description: id 与 path 均缺失；或目标不是顶层回收站条目（需先处理其父目录）
        '403':
          description: 无权限（CLOUDFS_PERMISSION_DENIED）
        '404':
          description: 回收站中不存在该条目
  /api/v1/cloudfs/trash/all:
    delete:
      tags:
      - CloudFS
      summary: 清空回收站
      description: 彻底删除当前 Space 回收站中的所有条目（含磁盘内容与元数据）。
      security:
      - BearerAuth: []
      - CookieAuth: []
      responses:
        '200':
          description: 已清空，返回清除条目数
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      purged:
                        type: integer
        '403':
          description: 非 Space 成员
  /api/v1/cloudfs/trash/restore:
    post:
      tags:
      - CloudFS
      summary: 还原回收站条目
      description: |
        清除 deleted_at 并把内容从 .trash 移回原位；文件夹会连同其已删除子树整体还原。
        原位置若已被占用则返回冲突（CLOUDFS_ALREADY_EXISTS），不发生任何修改。
        权限与删除一致：条目所有者本人或 Space 管理员。
      security:
      - BearerAuth: []
      - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: 条目 ID（与 path 二选一）
                path:
                  type: string
                  description: 条目路径（与 id 二选一）
      responses:
        '200':
          description: 已还原
        '400':
          description: id 与 path 均缺失；或目标不是顶层回收站条目（需先还原其父目录）
        '403':
          description: 无权限（CLOUDFS_PERMISSION_DENIED）
        '404':
          description: 回收站中不存在该条目
        '409':
          description: 原位置已被占用（CLOUDFS_ALREADY_EXISTS）
  /api/v1/cloudfs/usage:
    get:
      tags:
      - CloudFS
      summary: 获取存储用量与配额
      description: 返回当前 Space 的已用字节数、配额与使用百分比（已实现，此前遗漏登记）。
      security:
      - BearerAuth: []
      - CookieAuth: []
      responses:
        '200':
          description: 用量信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      used:
                        type: integer
                        format: int64
                      quota:
                        type: integer
                        format: int64
                      usedPercent:
                        type: number
        '403':
          description: 非 Space 成员
  # ═══════════════════════════════════════════════════════════
  # 云文件分享链接（DES-2026-0912-05）
  # 仅支持单文件分享；GET /api/v1/share/{token} 三条是**免鉴权**接口，
  # 其余三条挂在既有 cloudfs 鉴权组下。
  # ═══════════════════════════════════════════════════════════
...
  /api/v1/cloudfs/shares:
    get:
      tags:
      - CloudFS
      summary: 列出 Space 内的分享链接
      description: 返回非敏感字段；tokenHash / passwordHash 与明文 token 永不返回。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: fileId
        in: query
        required: false
        schema:
          type: string
        description: 只看某个文件的分享
      responses:
        '200':
          description: 分享列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            fileId:
                              type: string
                            fileName:
                              type: string
                            fileDeleted:
                              type: boolean
                            tokenPrefix:
                              type: string
                            requiresPassword:
                              type: boolean
                            expiresAt:
                              type: string
                              format: date-time
                            maxDownloads:
                              type: integer
                            downloadCount:
                              type: integer
                            revoked:
                              type: boolean
                            expired:
                              type: boolean
                            createdAt:
                              type: string
                              format: date-time
                            revokedAt:
                              type: string
                              format: date-time
                            lastAccessAt:
                              type: string
                              format: date-time
  /api/v1/cloudfs/share/{shareId}:
    delete:
      tags:
      - CloudFS
      summary: 撤销分享链接
      description: 创建者本人或 Space 管理员可撤销；撤销后 token 永久失效，不可恢复。
      security:
      - BearerAuth: []
      - CookieAuth: []
      parameters:
      - name: shareId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 已撤销
        '403':
          description: 无权撤销该分享
        '404':
          description: 分享不存在
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌认证，在Authorization头部中传递
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session Cookie认证

  # ═══════════════════════════════════════════════════════════
  # 请求参数定义 (Parameters)
  # ═══════════════════════════════════════════════════════════
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - code
      - message
      description: 统一错误响应格式
      properties:
        code:
          type: string
          description: 错误码
          example: USERNAME_EXISTS
        message:
          type: string
          description: 错误描述
          example: 用户名已被使用
        details:
          type: object
          nullable: true
          description: 额外错误详情
          example:
            field: username
            requirement: 至少3个字符
...